Refactor data for default login activity into subdir
We're gonna be refactoring most of that and folding it into our existing ViewModel setup

I appreciate the attempt at giving us a repository, but it doesn't really work very well with our current approach.
